Add WithQueryString Trait
