+++ src/Traits/Configuration/ComponentConfiguration.php
@@ -143,27 +143,6 @@ trait ComponentConfiguration
         return $this;
     }
 
-    public function setQueryStringStatus(bool $status): self
-    {
-        $this->queryStringStatus = $status;
-
-        return $this;
-    }
-
-    public function setQueryStringEnabled(): self
-    {
-        $this->setQueryStringStatus(true);
-
-        return $this;
-    }
-
-    public function setQueryStringDisabled(): self
-    {
-        $this->setQueryStringStatus(false);
-
-        return $this;
-    }
-
     public function setEagerLoadAllRelationsStatus(bool $status): self
     {
         $this->eagerLoadAllRelationsStatus = $status;
@@ -269,10 +248,4 @@ trait ComponentConfiguration
         return $this;
     }
 
-    public function setQueryStringAlias(string $queryStringAlias): self
-    {
-        $this->queryStringAlias = $queryStringAlias;
-
-        return $this;
-    }
 }
